fix: Fixing variable casting in services

+++ Back/Banco-de-Talentos/src/main/java/com/fractal/bancodetalentos/service/impl/MasterServiceImpl.java
@@ -30,7 +30,7 @@ public class MasterServiceImpl implements MasterService {
         List<LanguageResp> respList = new ArrayList<>();
         for (Object[] objects: result) {
             LanguageResp languageResp = new LanguageResp();
-            languageResp.setId((Integer) objects[0]);
+            languageResp.setId(((BigDecimal) objects[0]).intValue());
             languageResp.setName((String) objects[1]);
             languageResp.setCode((String) objects[2]);
             respList.add(languageResp);
@@ -46,7 +46,7 @@ public class MasterServiceImpl implements MasterService {
         List<RoleResp> roleRespList = new ArrayList<>();
         for (Object[] objects: list) {
             RoleResp roleResp = new RoleResp();
-            roleResp.setId((Integer) objects[0]);
+            roleResp.setId(((BigDecimal) objects[0]).intValue());
             roleResp.setName((String) objects[1]);
             roleResp.setCode((String) objects[2]);
             roleRespList.add(roleResp);
@@ -62,7 +62,7 @@ public class MasterServiceImpl implements MasterService {
         List<CurrenciesResp> currenciesResps = new ArrayList<>();
         for (Object[] objects: list) {
             CurrenciesResp currResp = new CurrenciesResp();
-            currResp.setId((Integer) objects[0]);
+            currResp.setId(((BigDecimal) objects[0]).intValue());
             currResp.setName((String) objects[1]);
             currResp.setCode((String) objects[2]);
             currenciesResps.add(currResp);
@@ -78,7 +78,7 @@ public class MasterServiceImpl implements MasterService {
         List<ProfileResp> profileResps = new ArrayList<>();
         for (Object[] objects: list) {
             ProfileResp profile = new ProfileResp();
-            profile.setId((Integer) objects[0]);
+            profile.setId(((BigDecimal) objects[0]).intValue());
             profile.setName((String) objects[1]);
             profile.setCode((String) objects[2]);
             profileResps.add(profile);
@@ -94,7 +94,7 @@ public class MasterServiceImpl implements MasterService {
         List<LangProficiencyResp> proficiencyResp = new ArrayList<>();
         for (Object[] objects: list) {
             LangProficiencyResp resp = new LangProficiencyResp();
-            resp.setId((Integer) objects[0]);
+            resp.setId(((BigDecimal) objects[0]).intValue());
             resp.setName((String) objects[1]);
             proficiencyResp.add(resp);
         }
@@ -109,7 +109,7 @@ public class MasterServiceImpl implements MasterService {
         List<CountryResp> respList = new ArrayList<>();
         for (Object[] objects: result) {
             CountryResp countryResp = new CountryResp();
-            countryResp.setId((Integer) objects[0]);
+            countryResp.setId(((BigDecimal) objects[0]).intValue());
             countryResp.setCountry((String) objects[1]);
             countryResp.setCode((String) objects[2]);
             respList.add(countryResp);
@@ -129,7 +129,7 @@ public class MasterServiceImpl implements MasterService {
         List<CityResp> respList = new ArrayList<>();
         for (Object[] objects: result) {
             CityResp cityResp = new CityResp();
-            cityResp.setId((Integer) objects[0]);
+            cityResp.setId(((BigDecimal) objects[0]).intValue());
             cityResp.setCity((String) objects[1]);
             respList.add(cityResp);
         }
